+++ tools/autograd/gen_annotated_fn_args.py
@@ -20,7 +20,8 @@ import textwrap
 
 from typing import Dict, List, Any
 
-from tools.codegen.gen import with_native_function, parse_native_yaml, FileManager
+from tools.codegen.gen import parse_native_yaml, FileManager
+from tools.codegen.context import with_native_function
 from tools.codegen.model import *
 import tools.codegen.api.python as python
 from .gen_python_functions import should_generate_py_binding, is_py_torch_function, is_py_nn_function, is_py_variable_method
