+++ tools/autograd/gen_variable_factories.py
@@ -8,7 +8,9 @@ from typing import Optional, List
 from tools.codegen.api.types import *
 import tools.codegen.api.cpp as cpp
 import tools.codegen.api.python as python
-from tools.codegen.gen import with_native_function, parse_native_yaml, FileManager, mapMaybe
+from tools.codegen.gen import parse_native_yaml, FileManager
+from tools.codegen.context import with_native_function
+from tools.codegen.utils import mapMaybe
 from tools.codegen.model import *
 
 OPTIONAL_TYPE_PATTERN = re.compile(r"c10::optional<(.+)>")
