+++ tools/autograd/gen_variable_type.py
@@ -36,7 +36,9 @@ from tools.codegen.api.types import *
 from tools.codegen.api.autograd import *
 import tools.codegen.api.cpp as cpp
 from tools.codegen.code_template import CodeTemplate
-from tools.codegen.gen import with_native_function, parse_native_yaml, FileManager, mapMaybe
+from tools.codegen.gen import parse_native_yaml, FileManager
+from tools.codegen.context import with_native_function
+from tools.codegen.utils import mapMaybe
 from tools.codegen.model import *
 from tools.codegen.selective_build.selector import SelectiveBuilder
 from typing import Callable, List, Optional, Sequence, Tuple, Union
