+++ tools/codegen/dest/__init__.py
@@ -1,0 +1,1 @@
+from .register_dispatch_key import RegisterDispatchKey as RegisterDispatchKey
